refactor: popupProvider 다시 추천받기 버튼 적용
